step2 : 지하철 구간 삭제 테스트코드 작성
